Moved code out of main.go
